EITrans::EITrans

EITrans::EITrans is the main function for ensemble forecast calibration.

EITrans(
  ens,
  ens_times,
  ens_flts,
  ens_times_train,
  ens_times_dev,
  ens_times_test,
  obs,
  left_deltas,
  right_deltas,
  infinity_estimator,
  multiplier,
  circular_ens = F,
  member_weights = NULL
)

Arguments

ens
A 4-dimensional array for ensemble forecasts. Dimensions should be [stations, times, lead times, members].

ens_times
A vector for ensemble forecast times.

ens_flts
A vector for ensemble forecast lead times.

ens_times_train
Training times from the ensemble forecast times.

ens_times_dev
Development times from the ensemble forecast times.

ens_times_test
Testing times from the ensemble forecast times.

obs
A 3-dimensional array for observations that correspond to the ensemble forecasts. Dimensions should be [stations, times, lead times].

left_deltas
A vector of left-edge deltas to experiment.

infinity_estimator
A vector of values to experiment for estimating the ensemble spread.

multiplier
A vector of values to experiment for adjusting the ensemble member offset.

circular_ens
Whether the ensemble forecast variable is circular.

member_weights
Weights for each ensemble members when finding similar historical ensemble forecasts.

rigth_deltas
A vector of right-edge deltas to experiment.

Value

A list with the calibrated ensemble forecasts and intermediate results.

Details

This function uses foreach parallel mechanism. Parallelization is handled by users creating the parallel backend. Please see examples. It is recommended to use doSNOW for handling backend.

Examples

if (FALSE) {
# Use doSNOW to launch parallel backend
library(doSNOW)
cl <- snow::makeCluster()
registerDoSNOW(cl)

# If you know a host file, you can also launch on remote server
nodefile <- Sys.getenv("PBS_NODEFILE")
nodes <- readLines(nodefile)
cl <- makeCluster(nodes, type = "SOCK")
registerDoSNOW(cl)

eitrans_results <- EITrans(
  ens = ens$analogs,
  ens_times = ens$test_times,
  ens_flts = ens$flts,

  ens_times_train = ens$test_times[1:(test_start - 366)],
  ens_times_dev = ens$test_times[(test_start - 365):(test_start - 1)],
  ens_times_test = ens$test_times[test_start:test_end],

  obs = ens$obs_aligned,

  left_deltas = seq(-0.046, 0.12, by = 0.002),
  right_deltas = seq(-0.02, 0.044, by = 0.002),
  infinity_estimator = seq(0.1, 0.5, by = 0.1),
  multiplier = seq(0.1, 1.1, by = 0.1))

# If you are using doSNOW
stopCluster(cl)
}