EITrans::heuristicFilter

EITrans::heuristicFilter is designed to improve the reliability of Analog Ensemble forecasts by using the leave-one-out method and the inverse transformation function. The filter works better with ensembles with large amount of members (usually at the level of a hunderd).

heuristicFilter(AnEn, config, final.ensemble.size,
  historical.similar.count = 1, LOO.test.size = 10, keep.table = T,
  keep.LOO.rank = T, keep.LOO = T, do.not.append = F,
  show.progress = T, silent = F, i.station = NA, i.test.day = NA,
  AnEn.LOO.test = NA, heuristic.ensemble.size = config$num_members,
  member.name = "analogs")

Arguments

AnEn
The AnEn results from generateAnalogs.

config
The configuration used by generateAnalogs.

final.ensemble.size
How many ensemble members to keep in the final ensemble forecasts.

historical.similar.count
The number of past similar forecasts to include when considering calculating the historical rank histogram.

LOO.test.size
The number of historical forecasts to carry out leave-one-out test.

keep.table
Whether to keep the summary table of similariy.

keep.LOO.rank
Whether to keep the verification rank histogram for the leave-one-out tests.

keep.LOO
Whether to keep the analogs and similarity results from the leave-one-out test if they are computed.

do.not.append
If set to True, results will not be appended to AnEn, and the filtered analog ensembles will be returned. At the same time, all keep* parameters become invalid.

show.progress
Whether to show the progress bar.

silent
Whether to be silent.

i.station
A vector of station indices of which analogs will be filtered.

i.test.day
A vector of test day indices of which analogs will be filtered.

AnEn.LOO.test
The precomputed AnEn results for leave-one-out tests. If not provided, LOO tests will be carried out at the spot. See details for instructions on precopmuting LOO tests.

heuristic.ensemble.size
The number of ensemble members to take from the LOO test results to construct the heuristic rank histogram. By default, it is the num_members member from input config.

member.name
The name of the member in AnEn to process. By default, it is analogs. But this parameter can be helpful if, for example, you want to process another member called analogs.alternative.

Details

To precompute AnEn.LOO.test, simply copy the values from search_times_compare to test_times_compare in the config which is passed to this function, and use the generateAnalogs to generate the AnEn.LOO.test. To save memory and space, you can optionally omit similarity information in the results by setting config$keep_similarity = F. The only required member is analogs.

You might wonder when you need to generate an updated version of AnEn.LOO.test. Usually, if you change the settings in config which would affect the search data, for example, search_times_compare, and the calculation of AnEn, for example, weights, you will need to regenerate the AnEn.LOO.test. Otherwise, you do not need to regenerate it. For example, you don't need to regenerate AnEn.LOO.test if you only change test_times_compare.